Refactored integration tests into specific suites and files rather than in a single file
