+++ integration-tests/suites/common/retry.go
@@ -1,4 +1,4 @@
-package integrationtests
+package common
 
 import (
 	"time"
@@ -12,12 +12,12 @@ const (
 type retryable = func() (string, error)
 
 // Simple retry in loop for commands produce only one string output and error
-func retry(f retryable) (output string, err error) {
+func Retry(f retryable) (output string, err error) {
 	for i := 0; i < max_retries; i++ {
 		output, err = f()
 		if err == nil {
 			return output, nil
-		} else if i != max_retries - 1 {
+		} else if i != max_retries-1 {
 			time.Sleep(retry_wait_time)
 		}
 	}
