+++ integration-tests/suites/common/utils.go
@@ -1,4 +1,4 @@
-package integrationtests
+package common
 
 import (
 	"strconv"
@@ -7,12 +7,12 @@ import (
 
 // containerShortID returns the first twelve character of a containerID
 // to match the shortened IDs returned by docker.
-func containerShortID(containerID string) string {
+func ContainerShortID(containerID string) string {
 	return containerID[0:12]
 }
 
 // quoteArgs will add quotes around any arguments require it for the shell.
-func quoteArgs(args []string) []string {
+func QuoteArgs(args []string) []string {
 	quotedArgs := []string{}
 	for _, arg := range args {
 		argQuoted := strconv.Quote(arg)
@@ -28,11 +28,11 @@ func quoteArgs(args []string) []string {
 // Generate the QA tag to be used for containers by attaching the contents of
 // the 'COLLECTOR_QA_TAG' environment variable if it exists. Return the base
 // tag as is otherwise.
-func getQATag(base_tag string) string {
+func GetQATag(base_tag string) string {
 	collector_qa_tag := ReadEnvVar("COLLECTOR_QA_TAG")
 
 	if collector_qa_tag != "" {
-		return base_tag + "-" + collector_qa_tag;
+		return base_tag + "-" + collector_qa_tag
 	}
 	return base_tag
 }
@@ -40,6 +40,6 @@ func getQATag(base_tag string) string {
 // Return the full image to be used for a QA container from a given image name
 // and a tag. The tag will be adjusted accordingly to the description of
 // 'getQaTag'
-func qaImage(image string, tag string) string {
-	return image + ":" + getQATag(tag);
+func QaImage(image string, tag string) string {
+	return image + ":" + GetQATag(tag)
 }
