+++ integration-tests/suites/missing_proc_scrape.go
@@ -1,0 +1,44 @@
+package suites
+
+import (
+	"os"
+	"time"
+
+	"github.com/stackrox/collector/integration-tests/suites/common"
+	"github.com/stretchr/testify/assert"
+)
+
+type MissingProcScrapeTestSuite struct {
+	IntegrationTestSuiteBase
+}
+
+func (s *MissingProcScrapeTestSuite) SetupSuite() {
+	_, err := os.Stat("/tmp/fake-proc")
+	assert.False(s.T(), os.IsNotExist(err), "Missing fake proc directory")
+
+	s.executor = common.NewExecutor()
+	s.collector = common.NewCollectorManager(s.executor, s.T().Name())
+
+	// Mount the fake proc directory created by 'create-fake-proc.sh'
+	s.collector.Mounts["/host/proc:ro"] = "/tmp/fake-proc"
+
+	err = s.collector.Setup()
+	s.Require().NoError(err)
+
+	err = s.collector.Launch()
+	s.Require().NoError(err)
+
+	time.Sleep(10 * time.Second)
+}
+
+func (s *MissingProcScrapeTestSuite) TestCollectorRunning() {
+	collectorRunning, err := s.executor.IsContainerRunning("collector")
+	s.Require().NoError(err)
+	assert.True(s.T(), collectorRunning, "Collector isn't running")
+}
+
+func (s *MissingProcScrapeTestSuite) TearDownSuite() {
+	err := s.collector.TearDown()
+	s.Require().NoError(err)
+	s.cleanupContainer([]string{"collector"})
+}
